Refactor: Separate tests that open the Preview dock from those that don't.           Break preview-independent tests out into a separate file.
